+++ .github/CHANGELOG.md
@@ -68,6 +68,7 @@
 1. [A32NX/GPWS] Implement simulated GPWC in rust systems - @lukecologne (luke)
 1. [A32NX/FWC] Improve CAB PR LO DIFF PR warning logic to avoid spurious warnings - @tracernz (Mike)
 1. [A32NX/FWS] Fixed stall warning not working - @tracernz (Mike)
+1. [A32NX/FWS] Add `IDG 1(2) DISCONNECTED` and `GEN 1(2) OFF` master cautions - @FozzieHi (fozzie)
 
 ## 0.14.0